Update plot_cm.py
Currently, the network border, illustrated through LineCollection, is drawn atop of the connectivity/adjacency matrix, which masks the first rows. 

I have put clip_on on False, so that the border can expand "out" of the matrix. I put zorder at -1, so it will be drawn underneath the matrices, and doubled the linewidth as half of the line will be covered by the matrix.